fix timed lock context management
